Extract Picker Iframe Initializer for separate class
